// uikit/theme/types.ts
export type Breakpoints = string[]

export type MediaQueries = {
  xs: string
  sm: string
  md: string
  lg: string
  xl: string
  nav: string
  xsMobile: string
  mobile: string
  tablet: string
  desktop: string
}

export type Grids = {
  mobile: string
  desktop: string
}

export type Spacing = number[]

export type Radii = {
  primary: string
  small: string
  default: string
  card: string
  circle: string
}

export type Shadows = {
  level1: string
  active: string
  success: string
  warning: string
  focus: string
  inset: string
}

export type Gradients = {
  primary: string
  reversed: string
}

export type Contrasts = {
  tertiary: string
}

export type Colors = {
  primary1: string
  primary2: string
  primary3: string
  primary4: string
  primary5: string
  primary6: string
  primary7: string
  primary8: string
  primary9: string
  primary10: string
  primary11: string
  primary12: string

  secondary1: string
  secondary2: string
  secondary3: string
  secondary4: string
  secondary5: string
  secondary6: string
  secondary7: string
  secondary8: string
  secondary9: string
  secondary10: string
  secondary11: string
  secondary12: string

  gray1: string
  gray2: string
  gray3: string
  gray4: string
  gray5: string
  gray6: string
  gray7: string
  gray8: string
  gray9: string
  gray10: string
  gray11: string
  gray12: string

  red1: string
  red2: string
  red3: string
  red4: string
  red5: string
  red6: string
  red7: string
  red8: string
  red9: string
  red10: string
  red11: string
  red12: string

  green1: string
  green2: string
  green3: string
  green4: string
  green5: string
  green6: string
  green7: string
  green8: string
  green9: string
  green10: string
  green11: string
  green12: string
  greenOpacity: string

  orange1: string
  orange2: string
  orange3: string
  orange4: string
  orange5: string
  orange6: string
  orange7: string
  orange8: string
  orange9: string
  orange10: string
  orange11: string
  orange12: string
  orangeOpacity: string

  white1: string
  white2: string
  white3: string
  white4: string
  white5: string
  white6: string
  white7: string
  white8: string
  white9: string
  white10: string
  white11: string
  gradients: Gradients
  contrast: Contrasts
  background: string
  text: string

  score: string
  purple: string
  inputBg: string

  turquoise: string
  violet: string
  blue: string
  brandGradient: string
  divider: string
  vividBlue: string
}

export type ZIndices = {
  dropdown: number
  modal: number
  searchResultsContainer: number
  modalOverlay: number
  walletSidePanelOverlay: number
  walletSidePanel: number
  navbar: number
  defaultDrawer: number
  filtersDrawer: number
  toast: number
}

export type Backgrounds = {
  primary: string
}